Create Detect Variable With NAs
As R is based on S, which is UNIX style software, it is mainly used for statistical analysis.  But in practice, most of our time has been spent on data manipulation, includes: check distribution, check missing values, and whether variable distribution make sense or not.  The question is that seems there is no function or package for this purpose. This may not be ideal, but I'll try to build a series functions within R to solve the problem.